iot/publisher/requirements.txt
paho-mqtt==1.6.1
requests==2.28.1
python-dotenv==0.21.0